pyproject.toml
[tool.poetry]
name = "pics-hosting"
version = "0.1.0"
description = "Хостинг для картинок телеграм"
authors = ["Adrian Makridenko <github.com/lyaguxafrog>"]
license = "MIT"
readme = "README.md"
packages = [{include = "pics_hosting"}]

[tool.poetry.dependencies]
python = "^3.10"
flask = "^2.3.0"
flask-admin = "^1.6.1"
python-dotenv = "^1.0.0"
flask-sqlalchemy = "^3.1.1"
flask-bootstrap = "^3.3.7.1"
sqlalchemy = "^2.0.21"
flask-security = "^3.0.0"
flask-login = "^0.6.2"
werkzeug = "^2.3.0"
flask-script = "^2.0.6"
pyotp = "^2.9.0"
pytelegrambotapi = "^4.14.0"
pillow = "^10.0.1"
flask-images = "^3.0.2"
apscheduler = "^3.10.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
